Tests/LibWeb/Text/expected/Editing/execCommand-is-a-noop.txt

execCommand("bold") returned false
Hello, world!<b>I'm totally bold rn</b>
queryCommandEnabled("bold") returned false
queryCommandIndeterm("bold") returned false
queryCommandState("bold") returned false
queryCommandSupported("bold") returned false
queryCommandValue("bold") returned ""
